Open-LLM-VTuber/src/open_llm_vtuber/config_manager/main.py
# config_manager/main.py
from pydantic import BaseModel, Field
from typing import Dict, ClassVar

from .system import SystemConfig
from .character import CharacterConfig
from .live import LiveConfig
from .computer_use import ComputerUseConfig
from .i18n import I18nMixin, Description


class Config(I18nMixin, BaseModel):
    """
    Main configuration for the application.
    """

    system_config: SystemConfig = Field(default=None, alias="system_config")
    character_config: CharacterConfig = Field(..., alias="character_config")
    live_config: LiveConfig = Field(default=LiveConfig(), alias="live_config")
    computer_use_config: ComputerUseConfig = Field(
        default=ComputerUseConfig(), alias="computer_use_config"
    )

    DESCRIPTIONS: ClassVar[Dict[str, Description]] = {
        "system_config": Description(
            en="System configuration settings", zh="系统配置设置"
        ),
        "character_config": Description(
            en="Character configuration settings", zh="角色配置设置"
        ),
        "live_config": Description(
            en="Live streaming platform integration settings", zh="直播平台集成设置"
        ),
        "computer_use_config": Description(
            en="Computer use agent configuration for AI-controlled automation",
            zh="AI控制自动化的计算机使用代理配置",
        ),
    }
